chore: move bank and hope git doesn't do anything stupid
