+++ src/main/java/com/abc/bank/Bank.java
@@ -1,4 +1,4 @@
-package com.abc;
+package com.abc.bank;
 
 import java.util.ArrayList;
 import java.util.List;
